[#9] Moved Application module to Grog.App to use Grog as API module
